+++ NEWS.md
@@ -1,5 +1,7 @@
 # development version 
 
+# mikropml 1.1.1
+
 - Fixed bugs related to grouping correlated features (#276, @kelly-sovacool).
     - Also, `group_correlated_features()` is now a user-facing function.